% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameCurrency.R
\name{valueGoods}
\alias{valueGoods}
\title{Calculates the values of virtual goods in terms of a currency}
\usage{
valueGoods(transactions)
}
\arguments{
\item{transactions}{A numeric matrix representing the transactions available in the game.
    Each row represents a transaction.
    The first column represents the cost of a transaction in some currency.
    If the currency is used up in the transaction, then the value should be negative.

    The remaining columns represent the quantity of virtual goods gained after the transaction,
    each column representing a specific virtual good.
    There should be at least as many transactions as there are virtual goods.}
}
\value{
Values of each virtual good in terms of the currency given in the matrix.
    If the matrix of transactions is over- or under-determined, then the values of goods
    are returned which minimizes the squared error between the actual and calculated total values
    of the transaction.
}
\description{
Calculates the values of virtual goods in terms of a currency
}
